allow local build and pypi upload

+++ upload.sh
@@ -4,6 +4,13 @@
 
 cd $(dirname $0) && set -xe
 
-./build.sh
-
-docker run --rm -it --env-file env.sh automon "$@"
+if [ "$@" == '--local' ]; then
+  source env.sh
+  python3 setup.py sdist bdist_wheel
+  twine check dist/*
+  python3 -m twine upload --repository $PYPI --repository-url $TWINE_REPOSITORY \
+    -u $TWINE_USERNAME -p $TWINE_PASSWORD --non-interactive --skip-existing dist/*
+else
+  ./build.sh
+  docker run --rm -it --env-file env.sh automon "$@"
+fi
